+++ test-e2e/cypress-src/integration/describe-enabled-test-disabled/a-file.js
@@ -1,5 +1,6 @@
 describe("List items", { failFast: { enabled: true } }, () => {
   before(() => {
+    cy.task("log", "Executing before hook");
     cy.visit("/");
   });