+++ test-e2e/cypress-src/support/index.js
@@ -1,9 +1,10 @@
 /* global Cypress */
 
+import addContext from "mochawesome/addContext";
 import "cypress-fail-fast";
 
 Cypress.on("test:after:run", (test) => {
   if (test.state === "failed") {
-    cy.task("log", "Executing test:after:run event in failed test");
+    addContext({ test }, "Executed test:after:run event in failed test");
   }
 });